Make the project pypy37-compatible.
Don't install mypy and black on pypy37, replace Literal usage with just
bool (it was only used in two places, both as Literal[True]), get rid of
scipy because it takes forever to build and was only used for one
function it wasn't even necessary for.

Also make docs an optional dependency instead of a dev one, for sanity.